Changed value representing regex flag
Changed the value representing the regex flag in the database from INTEGER to TINYINT

+++ commands/responses/removebadphrase.js
@@ -22,7 +22,7 @@ const Phrases = sequelize.define('phrases', {
 	response: Sequelize.TEXT,
 	delete: Sequelize.TINYINT,
 	timeout: Sequelize.INTEGER,
-	regex: Sequelize.INTEGER,
+	regex: Sequelize.TINYINT,
 });
 
 module.exports = {
